+++ installation/create-admin.php
@@ -1,30 +1,27 @@
 <?php
 
-/**
- * @var ObjectInstancier $objectInstancier
- */
-
 //TODO a mettre dans une Commande
 
 //Crée un admin (crée aussi le rôle admin et fixe les droits si il n'existe pas
 
+use Pastell\Service\TokenGenerator;
+
 require_once __DIR__ . '/../init.php';
 
-
 $login = get_argv(1);
-$password = get_argv(2);
-$email = get_argv(3);
+$email = get_argv(2);
 
-/** @var AdminControler $adminControler */
-$adminControler = $objectInstancier->getInstance(AdminControler::class);
+$adminControler = ObjectInstancierFactory::getObjetInstancier()->getInstance(AdminControler::class);
+$password = (new TokenGenerator())->generate();
 
 $result = $adminControler->createAdmin($login, $password, $email);
 
 if ($result) {
         echo "Administrateur $login créé avec succès\n";
+        echo "Mot de passe : $password\n";
 } else {
     echo $adminControler->getLastError()->getLastError() . "\n";
-    echo "Usage : {$argv[0]} login password email\n";
+    echo "Usage : {$argv[0]} login email\n";
     exit;
 }