+++ test/PHPUnit/api/DocumentAPIControllerTest.php
@@ -326,7 +326,12 @@ class DocumentAPIControllerTest extends PastellTestCase
         $roleSql->addDroit('readonly', 'entite:lecture');
         $roleSql->addDroit('readonly', 'test:lecture');
         $userId = $this->getObjectInstancier()->getInstance(UtilisateurCreator::class)
-            ->create('readonly', 'test', 'test', 'readonly@example.com');
+            ->create(
+                'readonly',
+                'D@iw3DDf41Nl$DXzMJL!Uc2Yo',
+                'D@iw3DDf41Nl$DXzMJL!Uc2Yo',
+                'readonly@example.com'
+            );
         $this->getObjectInstancier()->getInstance(RoleUtilisateur::class)->addRole($userId, 'readonly', self::ID_E_COL);
 
         $id_d = $this->createTestDocument();