+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - La base de données par défaut est mariadb version 10.9.3
 - Passage du jeu de caractère par défaut de la base de données à utf8mb4 (possibilité de mettre des caractères UTF8 étendus)
 - Actes et Helios automatique: Lorsque le traitement est terminé, il est possible de modifier le document pour cocher la case envoi_sae et faire l'action "Verser au SAE" (c'était déjà possible via le connecteur de purge) #1627
+- Vérification de la force du mot de passe sur les requêtes API de type POST ou PATCH
 
 ## Suppressions
 
@@ -109,7 +110,6 @@
 - Suppression des élements relatifs au centre de gestions sur la version standard. Il est possible de les remettre avec la feature flag CDGFeature.
 - Suppression de l'URL dans l'onglet de retour du SAE
 - Suppression du support MySQL. La seule base supportée est MariaDB.
-
 
 # [3.1.17] - 2022-09-12